# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_yourorg_secret-manager_internal_models.DriftEvent:
    properties:
      created_at:
        type: string
      detected_at:
        type: string
      diff:
        items:
          type: integer
        type: array
      git_version:
        items:
          type: integer
        type: array
      id:
        type: string
      k8s_version:
        items:
          type: integer
        type: array
      namespace:
        allOf:
        - $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.Namespace'
        description: Relationships
      namespace_id:
        type: string
      resolution_action:
        type: string
      resolved_at:
        type: string
      resolved_by:
        type: string
      resolver:
        $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.User'
      secret_name:
        type: string
    type: object
  github_com_yourorg_secret-manager_internal_models.Group:
    properties:
      azure_ad_gid:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.GroupPermission'
        type: array
      users:
        description: Relationships
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.User'
        type: array
    type: object
  github_com_yourorg_secret-manager_internal_models.GroupPermission:
    properties:
      created_at:
        type: string
      group:
        allOf:
        - $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.Group'
        description: Relationships
      group_id:
        type: string
      id:
        type: string
      namespace:
        $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.Namespace'
      namespace_id:
        type: string
      role:
        type: string
    type: object
  github_com_yourorg_secret-manager_internal_models.Namespace:
    properties:
      cluster:
        type: string
      created_at:
        type: string
      drift_events:
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.DriftEvent'
        type: array
      environment:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        description: Relationships
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.GroupPermission'
        type: array
      secret_drafts:
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.SecretDraft'
        type: array
    type: object
  github_com_yourorg_secret-manager_internal_models.SecretDraft:
    properties:
      commit_sha:
        type: string
      created_at:
        type: string
      data:
        items:
          type: integer
        type: array
      edited_at:
        type: string
      edited_by:
        type: string
      editor:
        $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.User'
      git_base_sha:
        type: string
      id:
        type: string
      namespace:
        allOf:
        - $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.Namespace'
        description: Relationships
      namespace_id:
        type: string
      published_at:
        type: string
      published_by:
        type: string
      publisher:
        $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.User'
      secret_name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  github_com_yourorg_secret-manager_internal_models.User:
    properties:
      azure_ad_oid:
        type: string
      created_at:
        type: string
      email:
        type: string
      groups:
        description: Relationships
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.Group'
        type: array
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_api.CreateSecretRequest:
    properties:
      data:
        additionalProperties: true
        type: object
      name:
        type: string
    type: object
...
  internal_api.UpdateSecretRequest:
    properties:
      data:
        additionalProperties: true
        type: object
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/yourorg/secret-manager/issues
  description: Production-ready secret management with drift detection and GitOps
    integration
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/yourorg/secret-manager
  title: Secret Manager API
  version: "1.0"
paths:
...
  /namespaces/{namespace}/secrets:
    get:
      consumes:
      - application/json
      description: Get all secrets in a namespace with optional status filter
      parameters:
      - description: Namespace ID (UUID)
        in: path
        name: namespace
        required: true
        type: string
      - description: Filter by status (draft, published, drifted)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of secrets
          schema:
            items:
              $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.SecretDraft'
            type: array
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Namespace not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List secrets in namespace
      tags:
      - secrets
    post:
      consumes:
      - application/json
      description: Creates a new secret draft in the specified namespace
      parameters:
      - description: Namespace ID (UUID)
        in: path
        name: namespace
        required: true
        type: string
      - description: Secret data
        in: body
        name: secret
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateSecretRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created secret
          schema:
            $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.SecretDraft'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Namespace not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Secret already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new secret
      tags:
      - secrets
  /namespaces/{namespace}/secrets/{name}:
    delete:
      consumes:
      - application/json
      description: 'Delete a secret from the database. Note: Deleting published secrets
        may cause drift.'
      parameters:
      - description: Namespace ID (UUID)
        in: path
        name: namespace
        required: true
        type: string
      - description: Secret name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Secret deleted successfully
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Secret not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete secret
      tags:
      - secrets
    get:
      consumes:
      - application/json
      description: Retrieve a specific secret by name from a namespace
      parameters:
      - description: Namespace ID (UUID)
        in: path
        name: namespace
        required: true
        type: string
      - description: Secret name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Secret details
          schema:
            $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.SecretDraft'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Secret not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get secret by name
      tags:
      - secrets
    put:
      consumes:
      - application/json
      description: Update secret data. Reverts published/drifted secrets to draft
        status automatically.
      parameters:
      - description: Namespace ID (UUID)
        in: path
        name: namespace
        required: true
        type: string
      - description: Secret name
        in: path
        name: name
        required: true
        type: string
      - description: Updated secret data
        in: body
        name: secret
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated secret
          schema:
            $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.SecretDraft'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Secret not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update secret
      tags:
      - secrets
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"